requirements.txt
pytest==7.1.3
selene>=2.0.0b5
requests~=2.28.1
python-dotenv~=0.20.0
voluptuous~=0.13.1
allure-pytest==2.10.0